Add special case to handle starting intermediate images
When executing the "run" step on an intermediate container,
buildrunner tries to pull the image, which will fail on.
Historically, the API returns a 200, with an error message,
so the pull appears (to buildrunner) to succeed.  With a
recent change to the docker engine, the API call has started
returning a 404, which buildrunner interprets as an error.

This change checks the name of the image requested against
our local images.  If we find an image locally whose hash
matches the name of the requested images, we use it and do
not issue a pull.