Add a pluggable "fetch" facility for retrieving files/artifacts: XENG-3746
The initial pluggable backends a sparse at the moment:

  * Github
  * File
  * HTTP/HTTPS (just stubbed - not implemented)

At some point an Artifactory backend should be added as well.

While this should be available throughout Buildrunner it is only used
in one case at the moment: supporting the ``redirect`` key/value in
``buildrunner.yaml``.  At some future point the ``fetch`` facility
should be available as a function in Jinja2 environments.