Additional PyPi/Artifactory unification for ``setup.py`` and ``requirements.txt``: XENG-3925
Due to setup/distutils managing setup requirements differently than
how ``pip`` processes ``requirements.txt`` some special handling is
necessary.  Even though distutils hands-off to ``pip``, it filters out
arguments that do not appear to be package names - this is why raw
lines from ``requirements.txt`` cannot be handed directly to the
``install_requires`` argument of ``setup()``.  This can be dealt with
by assisting and passing the ``--extra-index-url`` arguments to the
``dependency_links`` argument of ``setup()``.

There's another wrinkle: due to other differences in the way ``pip``
is invoked it only looks for packages in certain ways.  To address
this the parent PyPi package directory (``vcsinfo`` in this case) is
passed as well.  ``pip`` is able to look through a directory of the
raw package ``.tar.gz`` files and find an appropriate match.