Simple elimination of double output; handle None run status: XENG-3964, XENG-3961
It seems like console and log are redundant.  Ultimately one of these
should likely be removed as a parameter since it appears that there is
little distinction.  Another possibility is to allow them to be
exactly what they are: a console and a log file and that log should
never go to console and console should not go to log.  For now make
just a simple change by not passing both to the run() method.